- fixed issue with mouseover - fix after review
